backend/pyproject.toml
[project]
name = "omoi-os"
version = "0.1.0"
description = "OmoiOS - Multi-Agent Orchestration System"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [{name = "Kevin Hill"}]
requires-python = ">=3.12"
dependencies = [
    "openhands-sdk>=1.1.0",
    "openhands-tools>=1.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.34.0",
    "gunicorn>=21.0.0,<23",
    "pydantic>=2.7.0,<3",
    "pydantic-settings>=2.3.0,<3",
    "sqlalchemy>=2.0.0,<3",
    "psycopg[binary,pool]>=3.2.0,<4",
    "pgvector>=0.3.3,<1",
    "alembic>=1.13.0,<2",
    "redis>=5.0.0,<6",
    "whenever>=0.9.3",
    "pyyaml>=6.0.0,<7",
    "logfire>=4.16.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "fastembed>=0.2.0",
    "openai>=2.8.0",
    "numpy>=2.3.4",
    "anthropic>=0.73.0",
    "pydantic-ai>=1.39.0",
    "jinja2>=3.1.6",
    "jsonschema>=4.20.0,<5",
    "fastmcp>=2.14.0",
    "cryptography>=43.0.0,<44",
    "bcrypt>=4.0.0,<5",
    "python-jose[cryptography]>=3.3.0,<4",
    "python-multipart>=0.0.6,<1",
    "email-validator>=2.0.0,<3",
    "instructor>=1.13.0",
    "orjson>=3.11.4",
    "daytona>=0.112.0",
    "greenlet>=3.2.4",
    "openhands-workspace>=1.1.0",
    "openhands-agent-server>=1.1.0",
    "daytona-sdk>=0.119.0",
    "structlog>=24.0.0",
    "rich>=13.0.0",
    "claude-agent-sdk>=0.1.17",
    "claude-code-sdk>=0.0.25",
    "stripe>=7.0.0",
    "resend>=2.19.0",
    "taskiq-redis>=0.5.0",
    "taskiq>=0.12.0",
    "sentry-sdk[fastapi,sqlalchemy]>=2.0.0",
    "posthog>=3.0.0",
    "spec-sandbox",
]

[tool.uv.sources]
spec-sandbox = { workspace = true }

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
# Configure module discovery: omoi_os package is in the project root
module-name = "omoi_os"
module-root = ""

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "openhands-agent-server>=1.1.0",
    "openhands-workspace>=1.1.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.26.0",
]
test = [
    "pytest>=8.0.0,<9",
    "pytest-asyncio>=0.23.0,<1",
    "fakeredis>=2.20.0,<3",
    "pytest-cov>=4.1.0,<5",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.8.0",
    "aiosqlite>=0.19.0",
    "bcrypt>=4.0.0,<5.0.0",
    "greenlet>=3.0.0",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F"]
ignore = [
    "E402",  # module-import-not-at-top-of-file (117 intentional: load_dotenv/init before imports)
    "E501",  # line-too-long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["F401"]           # scripts use try/except imports to test availability
"migrations/*" = ["E722"]        # migrations use bare except in cleanup code
"omoi_os/workers/*" = ["F821"]   # forward references in type hints

